Fix the error with multiple imposition
The problem was how Content.Movable class handled its matrix
of transformation to be applied on top of the current transformation.
This matrix was being initialized from the current transformation,
effectively doubling it in the end (plus any transformations applied
to the Movable object.
With single imposition step (using one Movable instance) this was fine,
because the original transformation matrix was identity and as such
produced no wrong results when applied twice.